fix: wrap component that use useSearchParams with suspence to prevent error on build
